requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ffmpeg-python==0.2.0
future==1.0.0
idna==3.10
pycparser==2.22
requests==2.32.3
srt==3.5.3
tqdm==4.67.1
urllib3==2.3.0
vosk==0.3.44
websockets==15.0.1
